% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{\%within\%}
\alias{\%within\%}
\title{within}
\usage{
summits \%within\% twoCol
}
\arguments{
\item{summits}{A vector of numeric to be tested}

\item{towCol}{Two columns the values of the vectors should fall between}
}
\value{
a vector of logical the same length as summits
}
\description{
Determines if the values in summits fall between the columns of twoCol.
Note that the columns in twoCol must be integer values and that the order
has no impact.
}
\examples{
summits<-c(1,10,100,200)
body<-rbind(c(0,2),c(1,10),c(11,9),c(201,199))
summits \%within\% body
# or
do.call("\%within\%",list(summits,body))
# if there are two rows and not two columns the body
# will automaticaly be transposed
(summits \%within\% t(body)) == (summits \%within\% body)
}

