% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.r
\name{inbetween}
\alias{inbetween}
\title{within}
\usage{
inbetween(summits, starts, ends)
}
\description{
within
}
\examples{
summits<-c(1,10,100,200)
body<-rbind(c(0,2),c(1,10),c(11,9),c(201,199))
within(summits,body[,1],body[,2])
}

