% Generated by roxygen2 (4.0.1): do not edit by hand
\name{OLSonTransformation}
\alias{OLSonTransformation}
\title{OLS on transformed vectors}
\usage{
OLSonTransformation(mx, my, transformation)
}
\arguments{
\item{mx}{a \code{matrix} holding at least 4 columns for \code{i}, \code{j}, \code{t} and any number of \code{value}(s)}

\item{my}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}

\item{transformation}{tranforming function, e.g. \code{WithinTransformation2}}
}
\value{
numeric (beta)
}
\description{
OLS on transformed vectors
}
\examples{
\dontrun{
mx <- cbind(m$country1, m$country2, m$year, m$gattwto1)
my <- cbind(m$country1, m$country2, m$year, m$lnrgdp1)
OLSonTransformation(mx, my, WithinTransformation2)
OLSonTransformation(my, mx, WithinTransformation2)
}
}

