% Generated by roxygen2 (4.0.1): do not edit by hand
\name{varyingModel1}
\alias{varyingModel1}
\title{Varying model based on formula (6)}
\usage{
varyingModel1(mx, my)
}
\arguments{
\item{mx}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}

\item{my}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}
}
\value{
numeric value of OLS beta
}
\description{
Varying model based on formula (6)
}
\references{
László Balázsi, László Mátyás (2014): The Estimation of Varying Coefficients Multi-dimensional Panel Data Models. Formula (6).
}

