% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/basic-models.R
\name{WithinTransformation1}
\alias{WithinTransformation1}
\alias{WithinTransformation2}
\alias{WithinTransformation3}
\alias{WithinTransformation4}
\alias{WithinTransformation5}
\alias{WithinTransformation6}
\title{Within transformation}
\usage{
WithinTransformation1(i = m[, 1], j = m[, 2], t = m[, 3], value = m[,
  4:ncol(m)], m)
}
\arguments{
\item{i}{country 1 (vector)}

\item{j}{country 2 (vector)}

\item{t}{year (vector)}

\item{value}{vector to be transformed}

\item{m}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}
}
\value{
vector
}
\description{
Either \code{i}, \code{j}, \code{t} and \code{value} parameters or \code{m} data frame with all these four vectors are required.
}
\examples{
\dontrun{
str(WithinTransformation2(m=cbind(m$country1, m$country2, m$year, m$lnrgdp1)))
str(WithinTransformation2(m$country1, m$country2, m$year, m$lnrgdp1))
}
}
\references{
László Balázsi, László Mátyás and Tom Wansbeek (2014): The Estimation of Multi-dimensional Fixed Effects Panel Data Models. Formula (3), (6), (9), (11) and (13).
}

