% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unbalanced.R
\name{WithinTransformation1c}
\alias{WithinTransformation1c}
\alias{WithinTransformation2c}
\alias{WithinTransformation3c}
\alias{WithinTransformation4c}
\alias{WithinTransformation5c}
\alias{WithinTransformation6c}
\title{Within transformation with unbalanced data}
\usage{
WithinTransformation1c(i = m[, 1], j = m[, 2], t = m[, 3], value = m[,
  4:ncol(m)], m)
}
\arguments{
\item{i}{country 1 (vector)}

\item{j}{country 2 (vector)}

\item{t}{year (vector)}

\item{value}{vector to be transformed}

\item{m}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}
}
\value{
vector
}
\description{
Either \code{i}, \code{j}, \code{t} and \code{value} parameters or \code{m} data frame with all these four vectors are required.
}
\references{
László Balázsi, László Mátyás and Tom Wansbeek (2014): The Estimation of Multi-dimensional Fixed Effects Panel Data Models. Formula (22) and (23).
}

