% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varying.R
\name{varyingModel3}
\alias{varyingModel3}
\title{Varying model based on formula (12)}
\usage{
varyingModel3(mx, my, mz)
}
\arguments{
\item{mx}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}

\item{my}{a \code{matrix} holding 4 columns for \code{i}, \code{j}, \code{t} and \code{value}}
}
\value{
OLS beta
}
\description{
Varying model based on formula (12)
}
\references{
László Balázsi, László Mátyás (2014): The Estimation of Varying Coefficients Multi-dimensional Panel Data Models. Formula (12).
}

